requirements.txt
streamlit
langchain
langgraph
langchain-google-genai
python-dotenv
langchain_tavily
langchain_community
arxiv
wikipedia
youtube_search
langgraph-checkpoint-sqlite
langchain_experimental
stackapi
matplotlib 
ddgs
youtube-transcript-api
reportlab
langmem
fastapi
langchain-google-community
google-auth
google-auth-oauthlib
google-auth-httplib2 
google-api-python-client